perubahan

+++ pages/Sarpras/hapus_sarpras.php
@@ -1,7 +1,11 @@
 <?php
-include '../config/koneksi.php';
+include '../../config/koneksi.php';
 
 $id = $_GET['id'];
+
+if (!$conn) {
+    die('Database connection failed');
+}
 
 $query = mysqli_query($conn, "DELETE FROM sarpras WHERE id_sarpras='$id'");
